Vastly improve code modularity.
