+++ saylua/modules/explore/static-source/es6/Dungeon/Utils/game_logic.js
@@ -1,0 +1,47 @@
+import { OBSTRUCTIONS } from "./game_helpers";
+
+export function translatePlayerLocation(player, tileLayer, tileSet, direction) {
+  let p_x, p_y, g_x, g_y, goalCell, goalTile, tileType;
+
+  p_x = player.location.x;
+  p_y = player.location.y;
+
+  g_x = p_x;
+  g_y = p_y;
+
+  switch(direction) {
+    case 'up':
+      g_y = p_y - 1;
+      break;
+    case 'down':
+      g_y = p_y + 1;
+      break;
+    case 'left':
+      g_x = p_x - 1;
+      break;
+    case 'right':
+      g_x = p_x + 1;
+      break;
+  }
+
+  // Determine if we can visit this tile, theoretically.
+  try {
+    goalCell = tileLayer[g_y][g_x];
+  } catch (e) {
+    return player.location;
+  }
+
+  if (goalCell === undefined) {
+    return player.location;
+  }
+
+  // Determine if we can go there, physically.
+  goalTile = tileSet[goalCell.tile];
+  tileType = goalTile['type'];
+
+  if (OBSTRUCTIONS.indexOf(tileType) !== -1) {
+    return player.location;
+  }
+
+  return { "x": g_x, "y": g_y };
+}
