+++ saylua/wrappers.py
@@ -30,5 +30,3 @@ def admin_access_required(f):
         return f(*args, **kwargs)
 
     return decorated_function
-
-
